Calculate the proper parking orbit plane
The planner can now determine the plane (inclination and LAN) such, that
the ejection burn is purely prograde. This eliminates the normal
component of the ejection burn, saving delta-V. However, this means that
the planner no longer assumes an equatorial parking orbit, which makes
it slightly harder to use when the player's craft is already in an
equatorial orbit.

In addition, this adds an input field for a minimum inclination. This is
especially useful for custom solar systems (such as RSS), where the
launch site is not on the equator. The minimum inclination can then be
set to the latitude of the launch site, and the LAN of the proper
parking orbit is automatically determined (as well as the inclination
in case it needs to be higher than this).

I am pretty sure that I have the LAN calculation correct for all
combinations of inclination and departure velocity, but this is somewhat
hard to test within KSP since I cannot directly set the departure
velocity vector. I have created a Geogebra sketch to try and
validate the calculation (https://www.geogebra.org/m/xxvsgvvr).