Change the ejection angle display
This changes the ejection angle display, to make it useful for inclined
parking orbits. Instead of drawing the angle to the departure planet's
orbital prograde, it draws the angle between periapsis and the asymptote
of the departure orbit. Incidentally, it also changes the value in the
ejection details window to this angle.

In the process of investigating this, I found that quite a lot of
calculations are done for each entry in the porkchop plot which are then
immediately thrown away (see the LambertSolver.TransferDeltaV overload
without TransferDetails). It would be best to split these calculations
off, such that they are only performed for the one transfer that the
user selects (or that is automatically selected).