Ugly hack to make the parking orbit renderer work.
Creating a new class derived from OrbitTargetRenderer does not work -
the orbit lags behind the camera movement when panning. Therefore, we
need to use one of the built-in classes designed for rendering orbits.

The Setup method requires the contract to be non-null; however, in the
*_onUpdateCaption methods, a non-null contract will not work (it needs a
valid Agent, which is only set by an actual contract). We therefore
default-initialize a contract for the Setup method, and then immediately
set it to null before it is used.